% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_functions.R
\name{count_internal}
\alias{count_internal}
\title{Internal function to count the messages}
\usage{
count_internal(x, types)
}
\arguments{
\item{x}{a data.frame containing the message types and the counts}

\item{types}{a vector containing the types}
}
\value{
a numeric value of number of orders in x
}
\description{
Internal function to count the messages
}
\examples{
# Only used internally
}
\keyword{internal}
