% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_bytes}
\alias{format_bytes}
\title{Formats a number of bytes}
\usage{
format_bytes(x, digits = 2, unit_suffix = "B", base = 1000)
}
\arguments{
\item{x}{the value}

\item{digits}{the number of digits to display, default value is 2}

\item{unit_suffix}{the unit suffix, default value is 'B' (for bytes),
useful is also 'B/s' if you have read/write speeds}

\item{base}{the base for kilo, mega, ... definition, default is 1000}
}
\value{

}
\description{
Formats a number of bytes
}
\examples{
format_bytes(1234)
format_bytes(1234567890)
format_bytes(123456789012, unit_suffix = "iB", base = 1024)
}
