% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_exchange_from_filename}
\alias{get_exchange_from_filename}
\title{Returns the exchange from an ITCH-filename}
\usage{
get_exchange_from_filename(file)
}
\arguments{
\item{file}{a filename}
}
\value{
The exchange
}
\description{
Returns the exchange from an ITCH-filename
}
\examples{
get_exchange_from_filename("03302017.NASDAQ_ITCH50")
get_exchange_from_filename("20170130.BX_ITCH_50.gz")
get_exchange_from_filename("S030220-v50-bx.txt.gz")
}
