% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_data.R
\name{get_meta_data}
\alias{get_meta_data}
\title{Returns the meta data for the messages}
\usage{
get_meta_data()
}
\value{
a data.table with the information of the message-types
}
\description{
All information is handled according to the official ITCH 5.0
documentation as found here: 
\url{http://www.nasdaqtrader.com/content/technicalsupport/specifications/dataproducts/NQTVITCHSpecification.pdf}
}
\details{
\code{msg_type} the type of the message 
\code{msg_name} the official name of the message
\code{msg_group} the group the message belongs to
\code{doc_nr} the number of the message in the documentation
}
\examples{
get_meta_data()
}
