% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modifications.R
\name{get_modifications}
\alias{get_modifications}
\title{Retrieves the modifications of an ITCH-file}
\usage{
get_modifications(file, start_msg_count = 0, end_msg_count = 0,
  buffer_size = 1e+08, quiet = FALSE)
}
\arguments{
\item{file}{the path to the input file, either a gz-file or a plain-text file}

\item{start_msg_count}{the start count of the messages, defaults to 0}

\item{end_msg_count}{the end count of the messages, defaults to all messages}

\item{buffer_size}{the size of the buffer in bytes, defaults to 1e8 (100 MB), 
if you have a large amount of RAM, 1e9 (1GB) might be faster}

\item{quiet}{if TRUE, the status messages are supressed, defaults to FALSE}
}
\value{
a data.table containing the order modifications
}
\description{
If the file is too large to be loaded into the file at once,
you can specify different start_msg_count/end_msg_counts to load only some messages.
}
\examples{
\dontrun{
  raw_file <- "20170130.PSX_ITCH_50"
  get_modifications(raw_file)
  get_modifications(raw_file, quiet = TRUE)

  # load only the message 20, 21, 22 (index starts at 1)
  get_modifications(raw_file, startMsgCount = 20, endMsgCount = 22)
}

\dontrun{
  gz_file <- "20170130.PSX_ITCH_50.gz"
  get_modifications(gz_file)
  get_modifications(gz_file, quiet = T)
}
}
