% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orders.R
\name{get_orders}
\alias{get_orders}
\title{Retrieves the messages of an ITCH-file}
\usage{
get_orders(
  file,
  start_msg_count = 0,
  end_msg_count = -1,
  buffer_size = -1,
  quiet = FALSE,
  force_gunzip = FALSE,
  force_cleanup = FALSE
)
}
\arguments{
\item{file}{the path to the input file, either a gz-file or a plain-text file}

\item{start_msg_count}{the start count of the messages, defaults to 0, or a data.frame of msg_types and counts, as outputted by count_messages()}

\item{end_msg_count}{the end count of the messages, defaults to all messages}

\item{buffer_size}{the size of the buffer in bytes, defaults to 1e8 (100 MB), 
if you have a large amount of RAM, 1e9 (1GB) might be faster}

\item{quiet}{if TRUE, the status messages are supressed, defaults to FALSE}

\item{force_gunzip}{only applies if file is a gz-file and a file with the same (gunzipped) name already exists.
if set to TRUE, the existing file is overwritten. Default value is FALSE}

\item{force_cleanup}{only applies if file is a gz-file. If force_cleanup=TRUE, the gunzipped raw file will be deleted afterwards.}
}
\value{
a data.table containing the orders
}
\description{
If the file is too large to be loaded into the file at once,
you can specify different start_msg_count/end_msg_counts to load only some messages.
}
\examples{
\dontrun{
  raw_file <- "20170130.PSX_ITCH_50"
  get_orders(raw_file)
  get_orders(raw_file, quiet = TRUE)

  # load only the message 20, 21, 22 (index starts at 1)
  get_orders(raw_file, startMsgCount = 20, endMsgCount = 22)
}

\dontrun{
  gz_file <- "20170130.PSX_ITCH_50.gz"
  get_orders(gz_file)
  get_orders(gz_file, quiet = TRUE)
  
  msg_count <- count_messages(raw_file)
  get_orders(raw_file, msg_count)
}
}
