% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunzip_file.R
\name{gunzip_file}
\alias{gunzip_file}
\title{gunzips a binary file to a destination}
\usage{
gunzip_file(
  infile,
  outfile = gsub("\\\\.gz$", "", infile),
  buffer_size = min(4 * file.size(infile), 2e+09)
)
}
\arguments{
\item{infile}{the gzipped binary file}

\item{outfile}{a raw file}

\item{buffer_size}{the size of the buffer to read in at once, default is 4 times the file.size (max 2Gb).}
}
\value{
The filename of the unzipped file
}
\description{
gunzips a binary file to a destination
}
\examples{
\dontrun{
  file <- "20191230.BX_ITCH_50.gz"
  gunzip_file(file, "tmp")
}
}
