% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{open_itch_ftp}
\alias{open_itch_ftp}
\title{Opens the ITCH sample FTP page}
\usage{
open_itch_ftp()
}
\value{
the URL (invisible)
}
\description{
The FTP server can be found at \url{ftp://emi.nasdaq.com/ITCH/}.
}
\examples{
\dontrun{
open_itch_ftp()
}
}
