% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{open_itch_sample_server}
\alias{open_itch_sample_server}
\title{Opens the ITCH sample page}
\usage{
open_itch_sample_server()
}
\value{
the URL (invisible)
}
\description{
The server can be found at \url{https://emi.nasdaq.com/ITCH/Nasdaq\%20ITCH/}.
}
\examples{
\dontrun{
open_itch_sample_server()
}
}
