% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{open_itch_specification}
\alias{open_itch_specification}
\title{Opens the ITCH Specification URL}
\usage{
open_itch_specification()
}
\value{

}
\description{
The specifications can be found as a PDF \url{https://www.nasdaqtrader.com/content/technicalsupport/specifications/dataproducts/NQTVITCHspecification.pdf}.
}
\examples{
\dontrun{
open_itch_specification()
}
}
