% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_functions.R
\name{count_functions}
\alias{count_functions}
\alias{count_orders}
\alias{count_trades}
\alias{count_modifications}
\alias{count_system_events}
\alias{count_stock_directory}
\alias{count_trading_status}
\alias{count_reg_sho}
\alias{count_market_participant_states}
\alias{count_mwcb}
\alias{count_ipo}
\alias{count_luld}
\alias{count_noii}
\alias{count_rpii}
\title{Counts the number of messages in a group from a data.table of message counts}
\usage{
count_orders(x)

count_trades(x)

count_modifications(x)

count_system_events(x)

count_stock_directory(x)

count_trading_status(x)

count_reg_sho(x)

count_market_participant_states(x)

count_mwcb(x)

count_ipo(x)

count_luld(x)

count_noii(x)

count_rpii(x)
}
\arguments{
\item{x}{a file or a data.frame containing the message types and the counts}
}
\value{
a numeric value of number of orders in x
}
\description{

}
\details{
The count functions apply to all message groups.
}
\examples{
file <- system.file("extdata", "ex20101224.TEST_ITCH_50", package = "RITCH")
msg_count <- count_messages(file, quiet = TRUE)

count_orders(msg_count)

count_trades(msg_count)

count_modifications(msg_count)

count_system_events(msg_count)

count_stock_directory(msg_count)

count_trading_status(msg_count)

count_reg_sho(msg_count)

count_market_participant_states(msg_count)

count_mwcb(msg_count)

count_ipo(msg_count)

count_luld(msg_count)

count_noii(msg_count)

count_rpii(msg_count)
}
