% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_code.R
\name{download_locate_code}
\alias{download_locate_code}
\title{Downloads the locate code for a given date and exchange}
\usage{
download_locate_code(exchange, date, quiet = FALSE)
}
\arguments{
\item{exchange}{The exchange, either NASDAQ (equivalent to NDQ), BX, or PSX}

\item{date}{The date, should be of class Date. If not the value is converted using \code{as.Date}.}

\item{quiet}{If the download function should be quiet, default is FALSE.}
}
\value{
a data.table of the tickers, the respective locate codes, and the exchange/date information
}
\description{
The data is downloaded from NASDAQs FTP server, which can be found here 
\url{ftp://emi.nasdaq.com/ITCH/Stock_Locate_Codes/}
}
\examples{
\dontrun{
download_locate_code("BX", "2019-07-02")
download_locate_code(c("BX", "NDQ"), c("2019-07-02", "2019-07-03"))
}
}
