% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_date_from_filename}
\alias{get_date_from_filename}
\title{Returns the date from an ITCH-filename}
\usage{
get_date_from_filename(file)
}
\arguments{
\item{file}{a filename}
}
\value{
the date as fastPOSIXct
}
\description{
Returns the date from an ITCH-filename
}
\examples{
get_date_from_filename("03302017.NASDAQ_ITCH50")
get_date_from_filename("20170130.BX_ITCH_50.gz")
}
\keyword{internal}
