% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_samples.R
\name{list_sample_files}
\alias{list_sample_files}
\title{Returns a data.table of the sample files on the server}
\usage{
list_sample_files()
}
\value{
a data.table of the files
}
\description{
The Server can be found at \url{https://emi.nasdaq.com/ITCH/Nasdaq\%20ITCH/}
}
\examples{
\dontrun{
  list_sample_files()
}
}
